anfisa_for_friends/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-debug-toolbar==6.3.0
pytz==2026.1.post1
sqlparse==0.5.5
tzdata==2026.2
